server/requirements.txt
eventlet
flask
flask-cors
flask-socketio
gitpython
piADC
RPi.GPIO

-e git://github.com/adafruit/Adafruit-Motor-HAT-Python-Library.git#egg=Adafruit_MotorHAT
